requirements.txt
Flask
flask_wtf
Bootstrap_Flask
pillow
numpy
matplotlib
